logic removed from routes - controller created
